Now using explicit intermediate model with "through"
There is extra information such as question order in an exam, so extra
columns are needed in the linking table.

This complicates the exam form, because two models are involved.  Using
django-extra-views to combine an Exam form with inline formsets for
ExamQuestion.  This allows questions to be added to an exam, as well as
setting their order at the same time.

The usual problem arises, where swapping the order of questions causes
an invalid form.