rationalise preview/download views; make download work
need to throw a nice error page instead of JSON for the download view, but otherwise works fine.
We should also add some code in to check timestamps of output files against last-edited-times of objects, so that we don't needlessly recompile things.

I've abandoned the idea of including updated content data with the preview/download requests, so they're now GET requests which just trigger a compile-and-response, with no DB update.
Reasons:
 * you can put /exam/1/my-exam.zip in your address bar and get the .zip file. This might be good for sharing exams.
 * the autosave already does the saving. We can wait for an autosave request to finish before asking for the preview, to ensure the latest data are used.
 * for exams, I don't think the preview view was saving the question list

the preview view still returns JSON with a url in it, but ideally going to /exam/1/my-exam/preview/ should do the following:

* compile the preview
* redirect to /exam/1/my-exam/preview/index.html
* everything matching /exam/1/my-exam/preview/.* redirects to the appropriate static url

shouldn't be too hard.