Docs and settings tidying
Apparently, "globals" is a Python internal reserved word, so I've
changed it to something else, and moved the context processor stuff to a
different file, more in keeping with the way Django does it.

Updated settings.py.dist to reflect the installation instructions.

Added an Apache 2 config file, which requires almost no editing if
following the installation instructions.