fix exam edit in IE; replace file drag-and-drop with boring old button; fix tryLoad bug
The drag-and-drop file upload thing is far too much of a hassle to get to work on IE, so I've given up on it.
The exam editing page now works in IE.
There was a typo in Editor.tryLoad which meant it didn't work when the lower-case version of an attribute was present in the loaded data. Fixed.