Slug field should be the same length as name
The default size of the slug field was only 50 characters, whereas the
name field is 200.  If the name of the exam/question exceeded 50
characters Django complained that the data for the slug field was too
long.  Fixed by making the length of a slug field the same length as the
name field.