Allow LDAP authentication
This makes use of the django_auth_ldap backend.  If LDAP authentication
is enabled in settings.py, then when a user attempts to log in the
default behaviour is the following:
 * First check whether the user exists in LDAP.
  * If the user does exist in LDAP, but a django user does not exist, then
    create a django user, populating the User properties with LDAP
    attributes.  The password is not stored in the django database.
  * If the user does exist in LDAP, and a django user already exists,
    then authenticate against the LDAP password.
 * Second, check whether the user exists in the django database
   (manually created).
  * If the user does exist in the django database, authenticate and log
    in.
  * If the user does not exist in the django database, authentication
    fails and log in is denied.

If a user is created manually, with a manual password, then subsequently
also exists in LDAP, then either password will be valid, but attributes
from LDAP will take precedence (names, e-mail etc.).

If a user exists in the defined 'is_staff' or 'is_superuser' LDAP
groups, then that user is given the relevant privileges.