Comment SMTP settings; add default e-mail address
Django's default e-mail mechanism is SMTP.  Assuming a standard server
setup, none of the settings need to be set, so comment them all out, and
only uncomment if needed.

Also "auth" was added to the list of INSTALLED_APPS.  I'm guessing this
was actually meant to be "django.contrib.auth", which was already there.
Having auth there broke everything, so I've removed it.