exam edit: click to add questions to an exam instead of drag+drop
I didn't like the act of dragging a question from the search results into the question list.
Because of the way the sortable widget works, the original disappears from the search list while you're dragging it. I used to have it so the result gets added back in afterwards, but that doesn't feel nice.
Results now have a big plus button which adds the question to the bottom of the list when you click it. You can then drag it into the position you want.