Starting to redo navigation; users can star questions
Navigation and searching are becoming impossible. This commit starts the process of changing the question index page, so the main index has a few short lists of questions to go to and a search box, and the main listing is on a separate search page.
This also adds a "favourite questions" list per user, which is displayed on the question index.