custom tag model with 'official' field; tag admin actions
Tags can be marked as 'official', which means they will show up on the front page tag cloud (once I've made it)

This also adds two admin actions, "mark as official" and "merge tags". When tags are merged the tag which is most-used is kept. You can always go to that tag and change its name afterwards.