"insert image" modal can also insert iframes
If you upload a .html file as a resource, you can embed it in any content area by clicking the "insert/edit image" button and selecting it. There's some text next to html files in the picker saying they'll be embedded as iframes.
You can change the size of the iframe by selecting it and clicking the "insert/edit image" button again.
This is useful for HTML5 applets exported from GeoGebra.