Preambles
CSS and JS preambles let you do some things more conveniently.
The CSS preamble could be used to define some classes for common elements.
The JS preamble could be used to set up things like dynamic graphs.
The JS preamble is called as soon as the question is created, before variables are randomised. You can wait for the 'question-html-attached' event to do things with the HTML.