resources tab gives the actual URL
the name of the uploaded file might not be the same as the original, for example if it's a duplicate filename. The resources tab now gives the full relative URL of the image, as used in the img src attribute.

and the URL is now `resources/question-resources/<filename>`, instead of `resources/images/<filename>`, because not all resources are images.