Add "replace with gapfill" button for top-level parts
Parts which are not already gaps but could be (i.e., not information only) now have a "replace with a gap-fill" button next to the "copy part" button.
This replaces the part with a gap-fill whose first gap is the old part. The prompt of the gap-fill is copied over from the old part, with a [[0]] placeholder at the bottom to include the part-which-is-now-a-gap.

Fixes #215