Starting on edit history in question edit
The question edit page keeps track of changes made since the page was opened. You can move back and forward through the versions with the arrows in the top-right.
A version is a complete dump of the question. Restoring a version involves resetting the question entirely and loading in the data as if it's new.
This would be much nicer if it only made the changes necessary to get to the desired version, so knockout could keep as many of the DOM elements as possible, and the viewModel can keep unchanged objects.