add testing tools to variable tab
You can specify a condition on the set of question variables which must pass. The generator should re-run until the condition passes (or it's run too many times, and everything fails).

There's a button to test the condition, which generates variables as many times as possible in 3 seconds, and presents some stats on how often it succeeds, how many runs you'd expect to take, etc.