more control over variables test
You can specify how long (in seconds) to run the variables test for.
The stats summary gives a 95% confidence interval for Pr(success).
The stats summary is cleared as soon as any variable or function changes, or the condition changes.
The code checks for errors when compiling or applying the condition.
The "Running for X seconds..." message counts down, so you know it hasn't crashed.