integrate stamps with question editor
When you test run a question, a box pops up saying "is this question ready to use?"
This probably isn't the best way of prompting the user to stamp a question.

Stamps are now integrated with versions in the editing history timeline - you should be able to see whether the question was ready to use when a particular change was made.