only squash uncommented versions between stamps
Previously, a version without a comment was only shown if it was the latest version.
If there are two stamps with an uncommented change between them, it would be helpful to see the change.
So now, an uncommented version is only omitted from the timeline if the previous *displayed* item  in the timeline is also a version - this means that the latest version between stamps will always be shown.

This commit also adds placeholder text to the comment field.