show stamps in search results, and lots more
There's now a "status" column in the question search results, with the latest stamp. The title text on the stamp says who did it, and when.
Add another status option: "This should not be used".
Add a field current_stamp to EditorModel objects, set whenever a stamp is saved, so you can filter questions/exams on their current stamp.
The text in the "Filter questions by..." bit of the search form is a bit smaller.
You can ask only to see questions stamped as "ready to use".
In the question edit page, the editing history tab is now next to General, to make it more prominent. Might rename it.