evaluate just enough variables to test the condition before proceeding
Suppose you have a condition on the variables which only looks at variable C, and a load of variables which then depend on C.
It doesn't make sense to compute all of the variables before deciding if the value of C is acceptable.
This commit changes the computeVariables method so it just computes enough variables to evaluate the condition, and then evaluates the rest if the condition is satisfied.