most of the pull request mechanics work now
There's an "other versions" tab in the question edit page, which lists questions related to this one.
You can click on "compare" to go to a page where you can create a PR going either way.
Open PRs are listed underneath the list of other versions. You can accept or reject a pull request if you have editing rights for the question.

To do:
- when you create a PR and you can edit the destination question, it should just do the merge straight away.
- tidy up the "requests to merge" list.
- fill in the noneditable view.