admin icons are a bit more reusable
The CSS for admin icons (preview,copy, etc.) used to assume they were inside links and had no siblings, which is no good.
This adds a new template tag, admin_icon, which lets you stick an icon in anywhere, and name it by its function instead of the icon it uses.