m_n_x parts can have custom choices or answers lists; rejig interface
"Match choices with answers" parts can now have variable lists of choices and/or answers, given by JME expressions.

This also separates the interface into three tabs: one for choices, one for answers (both following the format for choices in the 1_n_2 and m_n_2 parts), and one for the marking matrix.
This gives you lots more space (and a rich editor!) to write the option text, and the marking matrix is a more compact grid.
The "custom marking matrix" option has moved from the marking options tab to the marking matrix tab.
This fixes #252