fix request.user in the top bar
apparently detailviews now set 'objectname' in the context data to self.object. When the view is on a user, like the profile view, that means that 'user' is the user being looked at, not the user making the request.
This changes the base.html template to set user=request.user, so we're definitely using the right object.