refactor exams and questions
this is going to take a long time!
Exam and Question have always had loads in common, so they extended an abstract EditorModel class.
All the associated models (highlights, access, etc.) were duplicated, one for each of Exam and Question.
In short, there was duplication all over the place.
Because I want to search on exams and questions together, I decided it's time to do this properly - have a concrete base model with the stuff held in common by questions and exams, and add smaller models with OneToOne relationships with the base, containing whatever extra information an exam or question needs.

So far, I've got a load of New* models, and data migrations which copy everything over from the old models to the new ones, including revisions.
On the models side, I just need to deal with pull requests.
Once that's done, I can get rid of the old models, and get rid of the New* prefix on the new models.

Django won't let me do that without rewriting everything else that relies on the old models first.
So I'm going to do that tomorrow!