update requirements.pip
A major release of django-notifications-hq introduced a breaking change.
This updates numbas/requirements.pip to specify exact, known-working, versions of each package.
I should update this every now and then. Going to investigate which packages are lacking python 3 support, with the aim of getting the editor to run in python 3.