pull request views work; copy to a different project
The views to create and close pull requests now work.
While I was doing this, I noticed that the "make a copy" link always made a copy in the same project. If you don't have editing access to that project, you're stuck.
So now there's an intermediate screen asking you which project you want to put the copy in. It also gives you a chance to change the name of the copy - we should get fewer "Bloggs's copy of X" names now, hopefully.
The CopyView is now abstracted to editoritem.views, eliminating some repeated code.