add exam intro/outro text
You can write some text which appears on the front page.
You can also give a list of (percentage, message) pairs to be shown on the results page. The message with the highest percentage less than or equal to the student's score is shown.

see numbas/Numbas#443 and numbas/Numbas#447