catch extensions or themes unavailable to compiler, before compiling
if you set settings.MEDIA_ROOT to a relative path, it's possible that
the Numbas compiler won't be able to find the file.
The CompileObject view now checks that the paths to extensions and the
selected theme exist before calling the compiler.
Extension and theme paths are made relative to the editor's working directory
(unless MEDIA_ROOT is set to an absolute path, as the documentation
 recommends)

This all works to catch the cases when:
 - an extension file that should exist but doesn't
 - MEDIA_ROOT is set to a path relative to the editor working directory

and throw a sensible error message rather than silently fail.

Thanks to Stuart Simpson for bringing this to my attention.