tinymce 4.5
This updates TinyMCE to version 4.5. The main bug that has been fixed
upstream is that noneditable items like gapfills don't break the
caret/selection any more.

As well as that, this commit fixes a few of the outstanding
TinyMCE-related issues: editors resize themselves 100ms after loading,
which should be enough time to get the correct size, and moving a part
up/down the list doesn't break the editor any more.

This also enables the "preview" button in TinyMCE, to give an instant
preview of how the content area will look, with variables subbed in. I
need to do more testing to see how this works with things like HTML
variables. It would be too complicated to run the question preamble or
render gaps properly - will have to add a disclaimer to the preview
window. This fixes #189.