fix showing/hiding of the gapfill button
The "insert gapfill" button is now always present in the toolbar for the
part prompt editor, but is disabled when the part type is not 'gapfill'.
TinyMCE does something *really* weird when you change the visibility of
the button after initialisation - I got a huge blank space between two
sections of the toolbar.

This also fixes the gap number prompt - when inserting a gap it picks
the smallest number that isn't already present, and it correctly picks
up the current gap's number when you double click to change it.

fixes #344