better numbering of questions in exam edit
When a question belongs to a group with "pick a random subset" or "all
quesitons, in random order", no number is shown.
For questions in groups with "all questions, in this order", the number
is shown, taking into account the number of questions that will be taken
from previous groups.

Fixes #341