only recreate questions in the basket/recent list if they've changed
Editor.mappedObservableArray now only runs the mapping function on
values when they don't match the previous value at that position in the
list.

This means that when the recent/basket question lists in the exam editor
are updated, it doesn't create new Question objects every time,
triggering a redraw.

Fixes #217