codemirror areas max-height: 90vh
The maximum height of code areas is now 90% of the height of the window.

I tried removing the max-height, but when you've got a *lot* of lines,
it can lead to a lot of scrolling to get to other parts of the
interface.

This way, the editor can just about fill the screen, and still leave a
bit of margin to give context.

I'm going to close #31 now, because I don't think I can do better than
this. As I see it, the difficulty was down to not being able to read
much of the code you're editing without scrolling, and this should make
that less of a problem.