lots of rejigging of the question editor
I want to make it more obvious how to proceed through the editor to
write a question.

The main tabs have been re-ordered:

* "Settings" is now "Metadata", and just above "Access", near the bottom
of the list.
* That means "Statement" is now the first tab, followed by "Parts",
"Variables" and "Advice".

The "marking" tab for most parts now has a few "must-set" options, then
a blank vertical space and the rest of the options under an "advanced
settings" header.

The allow fractions/display answer as a fraction fields disappear when a
precision restriction is applied.

There's help text at the top of more tabs, particularly 'scripts' and
'adaptive marking', explaining what they're for.

The "add a part/step/gap" buttons are dropdowns, so you have to pick the
type before the part is created, rather than creating a part with a
default type and then changing it.

The order of the parts in the dropdown now follows frequency of use a
bit more closely.

For JME parts, the "unexpected variable names" settings appear before
the less-useful length and string restrictions.

The choices tab comes before the marking tab in 1_n_2 and m_n_2 parts.