default status for items is 'draft'
To encourage authors to use the feedback labels, all exams and questions
begin with the status "draft", which can only be removed by adding a
feedback stamp.

The stamp at the top of the page and to the right of the title is now a
dropdown button, to apply feedback stamps directly. The dropdown also
has a published/not published status line, which takes you to the
access tab when you click on it.

The feedback dropdown in the left toolbar is no longer necessary.