improve the pager for editor sections
The idea is to use the pager to guide a novice author through the
process of creating a question/exam.

The pager to move between editor sections is now a knockout component,
`editor-pager`.

The "next section" link goes greeen when you've completed all the
pre-publication tasks on the current tab.

There's some text in the middle, which says "Before moving on, you
should <X>", where <X> is a description of an unfinished task, e.g.
"select a licence". When all tasks are complete, the text becomes "Move
on when you're ready".