Some more touches to the question group interface
see #349
The reordering handle is only shown when there's more than one group.
The "questions to use" form is only shown when there's more than one
question in the group.
The group names are always shown in the editor. "Show group names to
student" dictates whether the group names are shown when the exam is
run.