some improvements to "list of numbers" data type
Anything which passes Numbas.util.isNumber is now allowed: that includes
fractions. It might be worth expanding this to any JME expression that
evaluates to a number, because I can see people wanting to write
`sqrt(2)`. The point of the data type helpers is that they're useful for
people who haven't touched JME, so maybe that's not necessary.

Invalid entries get the 'has-error' class, so you can pick them out -
the error message "One of the values is not a number" is quite vague.