Marking test creates an instance of a question
The part marking unit tests used to just hook into Numbas.marking and
use their own code to set up the student's answer and compile feedback.

That doesn't work, or isn't easy, for things like gapfills, where you
need to set up several parts and let them refer to each other.

So this changes the unit tests to create an instance of the question as
a Numbas.Question object, i.e. the same thing the Numbas runtime uses,
and mark that.