starting work on custom part types
To make reusing marking algorithms easier, I think the correct thing to
do is to define new part types, which would act like the built-ins but
have user-defined marking algorithms and settings.

At the moment, you can create new part types through the web interface,
and questions keep track of which custom part types they use.

Need to make the editing interface work, render the fields for the
settings in the question editor, and actually implement it in the
runtime!