warnings for undefined variables in input options, and required note descriptions
If you refer in the definition of an input option to a variable which won't be defined in the scope when the
part is initialised, a warning is shown. I need to do the same for
marking notes, adding in the names of the notes.

The required notes now have descriptions, which can't be edited.