The student answer widget uses a real Part instance
I had been sort-of reusing the answer widgets between the editor and the
runtime, but using the editor's model of a part instead of an instance
of the real Part type from the runtime.

This meant a lot of code duplication, and lots of headaches in making
the two different objects work with the same widgets. It's much simpler
just to create an instance of the question for the "student answer"
marking test (we're doing this for the unit tests already anyway).

So now the widget you're shown in the editor should be the same one the
default theme produces, and it's created using a real Part instance, so
we don't have to duplicate any logic.