CPT input widgets work, as does gapfill.
The answer input widget for the marking test belonging to an instance of
a custom part type now works.
For gapfill parts, the marking test fills its answer in by looking at
the answers entered in the marking algorithm tabs of the child gaps.

The marking test doesn't run unless the question's variables have been
evaluated - this avoids a race condition when the page loads.