static input options for most CPT answer widget options
This adds a lot of options to each of the custom part type answer
widgets, roughly matching the relevant options from the built-in part
types.

These options, as well as the "input hint" field, have a "static?"
switch. When turned off, a JME field is shown instead, so you can
produce the value based on the part's settings.