avoid infinite loops when running marking tests
knockout can trigger the computed function which creates the question
instance for a marking test when something in the question looks like
it's changed, but the resulting JSON is the same.

This adds a check that the question JSON really is different to last
time a question instance was generated, and also copes with rapidly
changing question JSON causing multiple question instances being
generated before the first one has finished loading.