only edit one part at a time
The part editor now works like the variable editor: there's a list of
parts on the right, and the editing interface on the left only shows the
currently selected part.

Clicking on "add a part/gap/step" switches the main interface to a list
of part types; selecting one adds it in the appropriate place and
selects it to be edited.