keep track of contributors to exams and questions
This adds a Contributor model, linking users to items. References to
users not on the editor are allowed, for content that is uploaded from a
different editor instance. The user's name and a link to their profile
are saved.

The .exam source of questions and exams contains a 'contributors' list,
giving names and profile URLs of all contributors.

This fixes #401