fix size of variable list table cells
It was really annoying me that long strings would change the height of
the variable preview cell, making the "regenerate values" button move up
and down.

This fixes the height of the preview to one line-height, and makes it
take up as much width as possible. The button cells take up much less
width.

I've also collected all the status icons (lock, used in testing, source
of randomisation) on the left.