search bar repeats GET params, not fields in context form
The fix in #428 added a hidden input for every field in the 'form'
context variable. On the search page this is fine, but on other pages,
like the user profile settings page, this means that elements in
whatever other form is on the page gets repeated.

This commit changes the tempalte to add hidden inputs for all query
parameters in the current request. This might have unforeseen problems,
too: maybe the only safe option is to explicitly list the parameters
that need to be included.