fix #436 and #437
The search_url block defined in project_base.html wasn't overriding the one
in top-nav.html. I've just put an 'if in_project' statement in top-nav,
since that's the only other context for searches, anyway. Maybe there
should be a 'search_url' context variable in future, if more search
contexts are added.

I've changed the code that repeats search parameters in the top nav
form, so it only repeats fields from the search form, and excludes both
'query' and 'page'.