add summernote editor; prompts for item description
This adds Summernote (https://summernote.org), a lightweight WYSIWYG
editor, for use in fields that don't need TinyMCE. I don't have much
confidence in it, but it's nice and unobtrusive and shouldn't stall the
browser while it loads. At the moment I'm just using it in the
item description field.

This expands the help block below the description field to give some
prompts about what to write.

fixes #440