Gap insert dialog is a select box
Because gaps can have names, the gap selection dialog in the prompt
editor now shows a select box with the names of defined gaps, instead of
just a number input.

This fixes #460

Additionally, when the list of gaps changes or is reordered, the
references in the prompt are updated accordingly, so each gapfill still
refers to the same gap.