invert colours for basket icon when question has been added
A quick fix for numbas/Numbas#627: when a question is in the basket, the basket icon
has a background colour and the icon itself is rendered in white.

This makes the two states distinguishable even without colour.

This commit also updates the title attribute on the button to "this is
in your basket".