Editor interface for extension source
There's now an interface to edit extensions within the Numbas editor,
and the ability to create an extension entirely within the editor.

An extension created in the editor will get a .js file and a README.md,
with some minimal structure in each.

The editor tells you when you're editing the documentation (README.md,
README.html or just README), and when you're modifying the main script
file for the extension. When editing .js files, there's a link to help
with the Numbas JS API.