use the absolutely final marking feedback for the mark note
When compiling the feedback for the `mark` note in a part unit test, the
part is submitted and the list of feedback and credit for the whole part
are used.
This catches things like the min/max score limits.

This commit also moves the answer input box outside the tabs in the
marking test section, so you can look at variable values while filling
in answers.