Beginning a very basic LTI tool
The `numbas_editor_lti` app provides an LTI tool at the URL `/lti`. When
opened as an instructor, you can select an exam to show to students.
When opened as a student, the exam is shown with a simple SCORM API
wrapper. It saves attempt data to localstorage, then reports scores back
using the LTI outcomes service.
The only data saved on the server is:

* consumer keys and secrets, each linked to an editor user.
* for each LTI context, the id of the exam
* in each user's session, the data provided by the LTI launch.

At the moment, the instructor view just shows the most recently-viewed
exams, and doesn't make sure you're logged in.

The student view is a minimal working prototype.

There's no explanatory text anywhere, yet.