content areas don't load tinymce until clicked
TinyMCE takes a long time to initialise. Ideally, I'd use a WYSIWYG
editor which has a single instance across the whole page, but for now
we're stuck with TinyMCE.

To make moving between sections quicker, content areas now begin static
and only turn into TinyMCE editors when clicked. This is quite awkward
for the user, so this change might not stay.