set up a Site object in first_setup.py
You need to have a Site object with the domain name of the editor.
Django can't/won't work it out from the request.

This adds a question to first_setup.py which creates a Site object with
the right domain.

This also changes the registration view to use
sites.shortcuts.get_current_site.

see numbas/Numbas#671