lots of alternatives stuff
Alternatives are now called 'expected errors'. Gets the point across a
bit more clearly.

This rejigs the parts menu, since there are so many kinds of child part.
There are little headers, and the "add a gap/step/expected error"
buttons are in the menu instead of in the main part editor.

Expected errors don't load or save superfluous settings like prompt and
adaptive marking.