fix computation of variable names in JME correct answer
It used to sub in question variables, which didn't work properly when
the variables hadn't been evaluated yet. It doesn't matter what gets
subbed in - it's not going to add a free variable.

So instead, replace each set of curly braces with a constant, and then
look for free variables.

fixes #457