show stamps from individual users on editing history tab
We want to quickly check that everyone involved in the checking process
has checked a question.

The editing history tab now shows the latest stamp for each user who
has given a feedback stamp. Members of the project the item is in are
shown first, and non-members in a separate table.