notify by email when a stamp or comment is added
This also removes the 'watching_users' field on the EditorItem model and
replaces it with a method. It was hard to keep the field in sync with
the list of people who have access to the item, since project members
get access automatically.
We could add a "NotWatching" model to allow people to stop watching an
item that they'd automatically get notifications for.