Update ubuntu-web.rst (#1)
A working install process for Ubuntu 18.04.
The setfacl does not work as expected on EC2
Instead I used plain group permissions and group names

There has been some omissions in the described process which are now added. 
E.g. It was not indicated when to use which user as well.

Note: The ALLOWED HOST is not set by the script as intended and need to be set manually