Suggest to change type of custom functions
This adds some heavy-duty machinery to check for common problems with
custom functions.

The following are checked for:
    * return a type other than the one declared
    * return a NaN anywhere within a nested data structure.

If these happen while evaluating a variable, they're shown in a "Warnings" box under the variable's definition.
It tries to suggest a signature for each of the function's parameters.
If the suggested signature isn't exactly the current declared
signature, it offers to change it with a "Do it" button.

The suggested signatures prefer the 'number' data type over other
anything else that can cast to 'number'. If a list or dict has every
item with the same suggested signature 'sig', then it suggests 'list of
sig' or 'dict of sig'. This applies recursively, so it can suggest
things like 'list of list of number'.

This is a lot of work for a single problem - forgetting about the other
number types - but I think it's an easy mistake to make so worth having.

see #544