extract packages as part of pre/post_save signals
Previously, extracting extension/theme packages was done in the form
logic.

This was missed out when updating a package through Django's admin
interface, which uses its own form classes.

This commit adds receivers for the pre_save and post_save signals on
editable packages, to determine when the zipfile has changed and to
extract it when it has.

fixes #540