better notifications
The notifications alert now makes a request to
/notifications/unread_json which returns a JSON object containing the
HTML to display for the few most recent notifications, as well as the
total number of unread notifications and the timestamp of the last one.
This makes the notification updating much less shonky.

This also adds a page to view all of your notifications, paginated.