improvements to the theme editor
New themes now have a style.css file, with the default CSS variable
settings under :root, so they're easy to change.
New themes also have an inherit.txt referring to default, since almost
all themes will extend the default one.

When you add a file to the theme editor, it tries to put it in the right
directory, according to suffix: .html and .xslt go in templates, .js
goes in files/scripts, .md and .txt at the top, and everything else in
files/resources.

There's a drop-down box to change a file from the default theme: it
opens an editor for that file, and fills it with the contents of the
corresponding file in the default theme.