always use treeToJME to display variable values
tokenToDisplayString uses the locale's preferred notation, which can be
confusing when it uses a comma for the decimal point.
For example, a list with JME definition [1.2, 3.14159, 2.7] could be displayed as
    [1,2, 3,14159, 2,7]
This changes Editor.displayJMEValue to always use
Numbas.jme.display.treeToJME, which always uses the "plain" notation
style.

It also removes some duplicated code to do with variable strings that
are too long.